MAINT: make `_utils` a full module
